Projeto_equilibraai/requirements.txt
streamlit
pandas
numpy
psycopg2-binary
scikit-learn
tensorflow
transformers